pyproject.toml
[tool.poetry]
name = "ai-ideas-enhancer"
version = "0.1.0"
description = ""
authors = ["Mathew Hany <mathew.hanybb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.4.1"
requests = "^2.31.0"
click = "^8.1.7"


[tool.poetry.scripts]
idea = "ai_ideas_enhancer:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
